Integrate boost::program_options to accept options
We utilize the boost::program_options library to specify a few options
for users, most notably the -d (--details) option which specifies that
the program should display more detail on the stock quotes than the
standard amount.

We also create a simple usage message and notify the user that they must
enter a stock quote for the program to run.